+++ src/Utilities/Initializer.h
@@ -9,7 +9,7 @@
 #include "../Drawing/Graphics.h"
 #include "../Event/EventManager.h"
 #include "../Game/Entity.h"
-#include "../Game/Hitbox.h"
+#include "../Physics/Hitbox.h"
 #include "../Game/World.h"
 #include "../Sound/SoundManager.h"
